// new-admin-frontend/src/interfaces/storeSettings.ts
export default interface StoreSettings {
  config_cnpj?: string
  config_phone?: string
  config_cpf_checkout?: boolean
  config_hide_prices?: boolean
  config_store_city?: string
  config_store_number?: string
  config_pharmacist_crf?: string
  config_pharmacist_name?: string
  config_responsible_name?: string
  config_email?: string
  config_google_tag_manager_id?: string
  config_stock?: boolean
  config_show_celphone?: boolean
  config_farmaciasapp_integrationStatus?: string
  config_navbar_color?: string
  config_meta_keyword?: string
  config_meta_description?: string
  config_meta_title?: string
  config_cep?: string
  config_hotjar_id?: string
  config_owner?: string
  config_address?: string
  config_schema_markup?: string
  config_navbar_text_color?: string
  config_header_text_color?: string
  config_footer_text_color?: string
  config_whatsapp_button?: boolean
  config_withdraw?: boolean
  config_whatsapp_phone?: string
  config_tawk_embed?: string
  config_url?: string
  config_name?: string
  config_company_name?: string
  config_analytics_id?: string
  config_analytics_api_client_id?: boolean
  config_logo?: string
  gaview?: string
  ga_client_email?: string
  config_pixel_id?: string
  ga_private_key?: string
  config_ifood_client_secret?: string
  config_ifood_client_id?: string
  config_ifood_store_id?: string
  facebook_shopping?: string
  consulta_remedios?: string
  google_shopping?: string
  cote_facil?: string
  cliquefarma?: string
  farmacia_marketplace?: string
  config_cx_email?: string
  config_cx_id?: string
  config_shipping_courier?: boolean
  product_default_send_box?: number
  config_default_product_weight?: number
  config_default_product_height?: number
  config_default_product_length?: number
  config_default_product_width?: number
  config_best_shipping?: boolean
  config_best_shipping_client_id?: string
  config_best_shipping_client_secret?: string
  etlMergeableFields?: string[]
  config_secondary_color?: string
  config_new_layout?: boolean
  config_cta_color?: string
  config_social_login?: boolean
  config_stock_display?: boolean
  config_epharma_clientId?: string
  config_epharma_username?: string
  config_epharma_password?: string
  config_afe?: string
  config_local_delivery_rule?: string
  config_address_longitude?: number
  config_address_latitude?: number
  config_correios_free_from?: number
  config_pickup_in_store?: boolean
  config_best_shipping_free_from?: number
}